Remove Reset and ResetLong functions
They're a holdover from useing the `optreset` global variable to reset
the state of the option parser. However, since the entire parser's state
is maintained in a struct, instead of in global variables, there's less
need for the ability to reset things. Instead, you can just create a new
struct with a new state.

Fixes #2.